Rubber Ducks

Rubber Ducks

You can modify the name of both the file in the documentation

JuliaPackageTemplate.RubberDucks.rubber_ducks_in_earth — Function.

Computes the number of rubber ducks that would fill the volume of the Earth.

Assumes a spherical Earth, and liquid rubber ducks (to avoid the packing problem).

Arguments:

num_ruber::Real is the first function argument. [m^3]

r_earth::Real is the first function argument. Default: 6378000 [m]

Returns:

num_ducks::Float64 Number of rubber ducks that could

source

JuliaPackageTemplate.RubberDucks.RUBBER_DUCK_PER_SEC — Constant.

Number of rubber ducks produced per second in the entire world on average.

Units: ducks/sec

source