Submodule

Submodule

Write Markdown with inline comments to create documentation pages.

JuliaPackageTemplate.YourSubmodule.your_function — Function.

Write something about the purpose of the function up top.

Arguments:

arg_1::Integer Desribute first function argument. [input_units]

arg_2::Integer Desribute second function argument. [input_units]

Returns:

ret_1::Float64 Describe first return value. [input_units]

ret_2::Float64 Describe second return value. [input_units]

Notes:

Make specific comments about corner cases or things to know about usage here.

References:

List citations to references used in writing the function here

source